4.32.2-Implementado serviço REST para cidades

+++ src/main/java/br/mrcfood/infrastructure/repository/CidadeRepository.java
@@ -3,6 +3,7 @@ package br.mrcfood.infrastructure.repository;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Repository;
 
 import br.mrcfood.domain.entity.Cidade;
@@ -28,6 +29,9 @@ public class CidadeRepository {
 	
 	public void remover(Long id) {
 		Cidade cd = buscarPorId(id);
+		if(cd == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
 		repository.delete(cd);
 	}
